requirements.txt
httpx
python-dotenv
redis==4.0.2
logging